+++ package.json
@@ -41,7 +41,9 @@
   },
   "devDependencies": {
     "autoprefixer": "^6.0.2",
+    "babel-eslint": "^4.1.2",
     "browserify": "^10.2.1",
+    "eslint-plugin-react": "^3.3.2",
     "gulp": "^3.8.11",
     "gulp-concat": "^2.6.0",
     "gulp-flatten": "^0.2.0",
